이미지목록프로필선택

+++ behind-back/src/main/java/com/reboot/behind/data/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.reboot.behind.data.controller;
 
+import com.reboot.behind.config.security.auth.PrincipalDetails;
 import com.reboot.behind.data.dto.*;
 import com.reboot.behind.data.entity.User;
 import com.reboot.behind.service.UserService;
@@ -49,41 +50,72 @@ public class UserController {
             value = "디테일을 제외한 회원정보 수정"
             , notes = "디테일을 제외한 회원정보를 수정한다")
     @PatchMapping()
-    public ResponseEntity<UserResponseDto> changeUser(@RequestBody UserResponseDto userResponseDto){
+    public ResponseEntity<?> changeUser(@RequestBody UserResponseDto userResponseDto){
         System.out.println(userResponseDto);
         System.out.println("호호호호호호호");
-//        int id = Integer.parseInt(SecurityContextHolder.getContext().getAuthentication().getName());
-//        System.out.println(id);
-        UserResponseDto userChangeDto = userService.changeUser(userResponseDto);
+        PrincipalDetails pd = (PrincipalDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        int tokenid = pd.getUser().getId();
+        System.out.println(tokenid);
+        if (tokenid == userResponseDto.getId()){
+            UserResponseDto userChangeDto = userService.changeUser(userResponseDto);
+            return ResponseEntity.status(HttpStatus.OK).body(userChangeDto);
+        }
+        else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("권한이 없습니다");
+        }
 
-        return ResponseEntity.status(HttpStatus.OK).body(userChangeDto);
     }
     @ApiOperation(
             value = "회원 Detail 수정"
             , notes = "회원 Detail을 수정한다.")
     @PatchMapping("/detail")
-    public ResponseEntity<UserResponseDto> ChangeDetail(@RequestBody ChangeUserDetailDto changeUserDetailDto) throws Exception {
-        UserResponseDto userResponseDto = userService.ChangeDetail(changeUserDetailDto.getId(), changeUserDetailDto.getDetail());
+    public ResponseEntity<?> ChangeDetail(@RequestBody ChangeUserDetailDto changeUserDetailDto) throws Exception {
+        PrincipalDetails pd = (PrincipalDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        int tokenid = pd.getUser().getId();
+        System.out.println(tokenid);
+        if (tokenid==changeUserDetailDto.getId()){
+            UserResponseDto userResponseDto = userService.ChangeDetail(changeUserDetailDto.getId(), changeUserDetailDto.getDetail());
 
-        return ResponseEntity.status(HttpStatus.OK).body(userResponseDto);
+            return ResponseEntity.status(HttpStatus.OK).body(userResponseDto);
+        }
+        else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("권한이 없습니다");
+        }
+
     }
     @ApiOperation(
             value = "좋아요(팔로우) 좋아요 리스트에 추가"
             , notes = "좋아요(팔로우)를 누르면 팔로우 리스트에 추가한다")
     @PostMapping("/like")
     public ResponseEntity<String> createFollower(@RequestBody FollowerDto followerDto){
-        userService.saveFollower(followerDto);
+        PrincipalDetails pd = (PrincipalDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        int tokenid = pd.getUser().getId();
+        System.out.println(tokenid);
+        if (tokenid == followerDto.getUser()) {
+            userService.saveFollower(followerDto);
 
-        return ResponseEntity.status(HttpStatus.OK).body("팔로우 성공!");
+            return ResponseEntity.status(HttpStatus.OK).body("팔로우 성공!");
+        }
+        else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("권한이 없습니다");
+        }
     }
     @ApiOperation(
             value = "좋아요(팔로우) 삭제"
             , notes = "좋아요(팔로우)삭제 리스트에서 제거")
     @DeleteMapping("/like")
     public ResponseEntity<String> deleteFollower(@RequestBody FollowerDto followerDto) throws Exception {
-        userService.deleteFollower(followerDto);
+        PrincipalDetails pd = (PrincipalDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        int tokenid = pd.getUser().getId();
+        System.out.println(tokenid);
+        if (tokenid == followerDto.getUser()){
+            userService.deleteFollower(followerDto);
+            return ResponseEntity.status(HttpStatus.OK).body("팔로우 취소!");
+        }
+        else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("권한이 없습니다");
+        }
 
-        return ResponseEntity.status(HttpStatus.OK).body("팔로우 취소!");
     }
     @ApiOperation(
             value = "유저 검색"
@@ -111,18 +143,25 @@ public class UserController {
 
     @DeleteMapping()
     public ResponseEntity<String> deleteUser(Integer id) throws  Exception{
-        userService.deleteUser(id);
-        return ResponseEntity.status(HttpStatus.OK).body("유저 삭제 완료!");
+        PrincipalDetails pd = (PrincipalDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        int tokenid = pd.getUser().getId();
+        if (tokenid==id){
+            userService.deleteUser(id);
+            return ResponseEntity.status(HttpStatus.OK).body("유저 삭제 완료!");
+        }
+        else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("권한이 없습니다");
+        }
     }
 
     @ApiOperation(
             value = "유저 이미지 조회"
             , notes = "유저의 이미지 목록을 조회한다")
 
-    @GetMapping("/img")
+    @GetMapping("/images")
     public ResponseEntity<?> getUserImages(Integer id){
 //        userService.deleteUser(id);
-        List<ImageResponseDto> userImages = userService.getUserImage(id);
+        List<String> userImages = userService.getUserImage(id);
         return ResponseEntity.status(HttpStatus.OK).body(userImages);
     }
 
@@ -130,11 +169,17 @@ public class UserController {
             value = "유저 프로필 저장"
             , notes = "유저의 프로필을 저장한다.")
 
-    @PatchMapping("/img")
+    @PatchMapping("/images")
     public ResponseEntity<String> selectProfileImage(Integer id, String image){
-        userService.saveProfile(id,image);
-//        List<ImageResponseDto> userImages = userService.getUserImage(image);
-        return ResponseEntity.status(HttpStatus.OK).body("프로필 등록 완료!");
+        PrincipalDetails pd = (PrincipalDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        int tokenid = pd.getUser().getId();
+        if (tokenid == id){
+            userService.saveProfile(id,image);
+            return ResponseEntity.status(HttpStatus.OK).body("프로필 등록 완료!");
+        }
+        else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("권한이 없습니다");
+        }
     }
 
 }
